update festival-unidadDeVenta
ahora funciona correctamente, en el caso de uso del test caso de uso pedido se le asigna localmente el id a festival, ya no es necesario por eso da error, por ultimo el caso de uso unidad de venta le falta en los constructores de foodtruck y puesto desarmable  asociarlos con un festival

+++ src/datos/Festival.java
@@ -6,7 +6,7 @@ import java.util.Set;
 
 public class Festival {
 
-	private long idFestival;
+	private int idFestival;
 	private String nombre;
 	private String temporada;
 	private LocalDate fechaInicio;
@@ -40,7 +40,7 @@ public class Festival {
 	public long getIdFestival() {
 		return idFestival;
 	}
-	protected void setIdFestival(long idFestival) {
+	protected void setIdFestival(int idFestival) {
 		this.idFestival = idFestival;
 	}
 	public String getNombre() {
